clean  up interface class and prediction models
