+++ Map.py
@@ -69,6 +69,7 @@ def CreateInitialMap(df):
             
     # ctx.add_basemap(ax, url=ctx.providers.Stamen.TonerLite, zoom=12)
     ax.set_axis_off()
+    # plt.show()
     return plt
 
 
@@ -109,7 +110,7 @@ def PathMap(sp, df):
           if sp.__contains__(a.Point):
               plt.text(a.geometry.x, a.geometry.y, a.Point)
               
-    ctx.add_basemap(ax, url=ctx.providers.Stamen.TonerLite, zoom=12)
+    # ctx.add_basemap(ax, url=ctx.providers.Stamen.TonerLite, zoom=12)
     ax.set_axis_off()
     plt.show()
