+++ data/my_test.csv
@@ -1,29 +1,17 @@
 TIME,SCATS,DAY,N,E,S,W,COUNT
-730,3682,0,1,0,0,0,229
-845,3682,0,1,0,0,0,215
-1015,3682,0,1,0,0,0,130
-1100,3682,0,1,0,0,0,113
-1130,3682,0,1,0,0,0,145
-1200,3682,0,1,0,0,0,107
-745,3682,0,1,0,0,0,224
-915,3682,0,1,0,0,0,190
-930,3682,0,1,0,0,0,160
-1000,3682,0,1,0,0,0,163
-1030,3682,0,1,0,0,0,129
-1100,3682,0,1,0,0,0,112
-1145,3682,0,1,0,0,0,121
-830,3682,0,1,0,0,0,219
-915,3682,0,1,0,0,0,173
-1000,3682,0,1,0,0,0,151
-1015,3682,0,1,0,0,0,123
-1045,3682,0,1,0,0,0,111
-1100,3682,0,1,0,0,0,132
-715,3682,0,1,0,0,0,176
-845,3682,0,1,0,0,0,198
-1100,3682,0,1,0,0,0,135
-1130,3682,0,1,0,0,0,143
-1215,3682,0,1,0,0,0,155
-1245,3682,0,1,0,0,0,120
-1300,3682,0,1,0,0,0,142
-730,3682,0,1,0,0,0,229
-1015,3682,0,1,0,0,0,139
+915,2000,1,1,0,0,0,257
+1030,2000,1,1,0,0,0,210
+1045,2000,1,1,0,0,0,212
+1215,2000,1,1,0,0,0,212
+1000,2000,1,1,0,0,0,239
+1200,2000,1,1,0,0,0,168
+915,2000,1,1,0,0,0,255
+930,2000,1,1,0,0,0,251
+945,2000,1,1,0,0,0,236
+1115,2000,1,1,0,0,0,203
+1130,2000,1,1,0,0,0,213
+1200,2000,1,1,0,0,0,221
+1300,2000,1,1,0,0,0,229
+1200,2000,1,1,0,0,0,196
+1015,2000,1,1,0,0,0,221
+1145,2000,1,1,0,0,0,222
