+++ data/my_train.csv
@@ -1,98 +1,70 @@
 TIME,SCATS,DAY,N,E,S,W,COUNT
-700,3682,0,1,0,0,0,154
-715,3682,0,1,0,0,0,196
-745,3682,0,1,0,0,0,271
-800,3682,0,1,0,0,0,254
-815,3682,0,1,0,0,0,249
-830,3682,0,1,0,0,0,231
-900,3682,0,1,0,0,0,166
-915,3682,0,1,0,0,0,167
-930,3682,0,1,0,0,0,146
-945,3682,0,1,0,0,0,139
-1000,3682,0,1,0,0,0,142
-1030,3682,0,1,0,0,0,135
-1045,3682,0,1,0,0,0,133
-1115,3682,0,1,0,0,0,129
-1145,3682,0,1,0,0,0,141
-1215,3682,0,1,0,0,0,133
-1230,3682,0,1,0,0,0,104
-1245,3682,0,1,0,0,0,134
-1300,3682,0,1,0,0,0,109
-700,3682,0,1,0,0,0,161
-715,3682,0,1,0,0,0,224
-730,3682,0,1,0,0,0,235
-800,3682,0,1,0,0,0,269
-815,3682,0,1,0,0,0,244
-830,3682,0,1,0,0,0,268
-845,3682,0,1,0,0,0,241
-900,3682,0,1,0,0,0,192
-945,3682,0,1,0,0,0,164
-1015,3682,0,1,0,0,0,134
-1045,3682,0,1,0,0,0,129
-1115,3682,0,1,0,0,0,140
-1130,3682,0,1,0,0,0,128
-1200,3682,0,1,0,0,0,123
-1215,3682,0,1,0,0,0,118
-1230,3682,0,1,0,0,0,142
-1245,3682,0,1,0,0,0,130
-1300,3682,0,1,0,0,0,111
-700,3682,0,1,0,0,0,167
-715,3682,0,1,0,0,0,197
-730,3682,0,1,0,0,0,227
-745,3682,0,1,0,0,0,246
-800,3682,0,1,0,0,0,256
-815,3682,0,1,0,0,0,272
-845,3682,0,1,0,0,0,206
-900,3682,0,1,0,0,0,176
-930,3682,0,1,0,0,0,168
-945,3682,0,1,0,0,0,140
-1030,3682,0,1,0,0,0,149
-1115,3682,0,1,0,0,0,144
-1130,3682,0,1,0,0,0,119
-1145,3682,0,1,0,0,0,128
-1200,3682,0,1,0,0,0,120
-1215,3682,0,1,0,0,0,125
-1230,3682,0,1,0,0,0,152
-1245,3682,0,1,0,0,0,105
-1300,3682,0,1,0,0,0,110
-700,3682,0,1,0,0,0,175
-730,3682,0,1,0,0,0,231
-745,3682,0,1,0,0,0,223
-800,3682,0,1,0,0,0,265
-815,3682,0,1,0,0,0,261
-830,3682,0,1,0,0,0,193
-900,3682,0,1,0,0,0,172
-915,3682,0,1,0,0,0,182
-930,3682,0,1,0,0,0,155
-945,3682,0,1,0,0,0,132
-1000,3682,0,1,0,0,0,137
-1015,3682,0,1,0,0,0,149
-1030,3682,0,1,0,0,0,141
-1045,3682,0,1,0,0,0,106
-1115,3682,0,1,0,0,0,125
-1145,3682,0,1,0,0,0,113
-1200,3682,0,1,0,0,0,140
-1230,3682,0,1,0,0,0,109
-700,3682,0,1,0,0,0,148
-715,3682,0,1,0,0,0,167
-745,3682,0,1,0,0,0,249
-800,3682,0,1,0,0,0,236
-815,3682,0,1,0,0,0,279
-830,3682,0,1,0,0,0,257
-845,3682,0,1,0,0,0,207
-900,3682,0,1,0,0,0,189
-915,3682,0,1,0,0,0,183
-930,3682,0,1,0,0,0,141
-945,3682,0,1,0,0,0,164
-1000,3682,0,1,0,0,0,139
-1030,3682,0,1,0,0,0,140
-1045,3682,0,1,0,0,0,143
-1100,3682,0,1,0,0,0,143
-1115,3682,0,1,0,0,0,117
-1130,3682,0,1,0,0,0,132
-1145,3682,0,1,0,0,0,124
-1200,3682,0,1,0,0,0,131
-1215,3682,0,1,0,0,0,145
-1230,3682,0,1,0,0,0,151
-1245,3682,0,1,0,0,0,132
-1300,3682,0,1,0,0,0,145
+900,2000,1,1,0,0,0,281
+930,2000,1,1,0,0,0,260
+945,2000,1,1,0,0,0,240
+1000,2000,1,1,0,0,0,222
+1015,2000,1,1,0,0,0,202
+1100,2000,1,1,0,0,0,201
+1115,2000,1,1,0,0,0,186
+1130,2000,1,1,0,0,0,205
+1145,2000,1,1,0,0,0,206
+1200,2000,1,1,0,0,0,200
+1230,2000,1,1,0,0,0,189
+1245,2000,1,1,0,0,0,186
+1300,2000,1,1,0,0,0,167
+900,2000,1,1,0,0,0,268
+915,2000,1,1,0,0,0,254
+930,2000,1,1,0,0,0,254
+945,2000,1,1,0,0,0,268
+1015,2000,1,1,0,0,0,225
+1030,2000,1,1,0,0,0,240
+1045,2000,1,1,0,0,0,193
+1100,2000,1,1,0,0,0,190
+1115,2000,1,1,0,0,0,195
+1130,2000,1,1,0,0,0,220
+1145,2000,1,1,0,0,0,211
+1215,2000,1,1,0,0,0,213
+1230,2000,1,1,0,0,0,199
+1245,2000,1,1,0,0,0,191
+1300,2000,1,1,0,0,0,205
+900,2000,1,1,0,0,0,258
+1000,2000,1,1,0,0,0,202
+1015,2000,1,1,0,0,0,208
+1030,2000,1,1,0,0,0,220
+1045,2000,1,1,0,0,0,193
+1100,2000,1,1,0,0,0,174
+1145,2000,1,1,0,0,0,203
+1215,2000,1,1,0,0,0,212
+1230,2000,1,1,0,0,0,196
+1245,2000,1,1,0,0,0,210
+900,2000,1,1,0,0,0,309
+915,2000,1,1,0,0,0,218
+930,2000,1,1,0,0,0,259
+945,2000,1,1,0,0,0,236
+1000,2000,1,1,0,0,0,193
+1015,2000,1,1,0,0,0,238
+1030,2000,1,1,0,0,0,205
+1045,2000,1,1,0,0,0,203
+1100,2000,1,1,0,0,0,207
+1115,2000,1,1,0,0,0,234
+1130,2000,1,1,0,0,0,224
+1145,2000,1,1,0,0,0,217
+1215,2000,1,1,0,0,0,218
+1230,2000,1,1,0,0,0,226
+1245,2000,1,1,0,0,0,191
+1300,2000,1,1,0,0,0,212
+900,2000,1,1,0,0,0,280
+915,2000,1,1,0,0,0,241
+930,2000,1,1,0,0,0,270
+945,2000,1,1,0,0,0,250
+1000,2000,1,1,0,0,0,208
+1030,2000,1,1,0,0,0,224
+1045,2000,1,1,0,0,0,203
+1100,2000,1,1,0,0,0,187
+1115,2000,1,1,0,0,0,212
+1130,2000,1,1,0,0,0,208
+1200,2000,1,1,0,0,0,185
+1215,2000,1,1,0,0,0,163
+1230,2000,1,1,0,0,0,220
+1245,2000,1,1,0,0,0,193
+1300,2000,1,1,0,0,0,199
